+++ src/flop.sol
@@ -81,7 +81,7 @@ contract Flopper is DSNote {
     }
 
     // --- Math ---
-    function add(uint x, uint y) internal pure returns (uint z) {
+    function add(uint48 x, uint48 y) internal pure returns (uint48 z) {
         z = x + y;
         require(z >= x);
     }
@@ -100,7 +100,7 @@ contract Flopper is DSNote {
         bids[id].bid = bid;
         bids[id].lot = lot;
         bids[id].guy = gal;
-        bids[id].end = uint48(add(now, tau));
+        bids[id].end = add(uint48(now), tau);
 
         emit Kick(id, lot, bid, gal, bids[id].end);
     }
@@ -117,7 +117,7 @@ contract Flopper is DSNote {
 
         bids[id].guy = msg.sender;
         bids[id].lot = lot;
-        bids[id].tic = uint48(add(now, ttl));
+        bids[id].tic = add(uint48(now), ttl);
     }
     function deal(uint id) public note {
         require(bids[id].tic < now && bids[id].tic != 0 ||
